컴포넌트 분리, 로딩 스피너를 스켈레톤으로 변경
